ai-services/requirement.txt
fastapi
uvicorn[standard]
langchain
faiss-cpu
pydantic
openai
python-dotenv
langchain-community
pypdf
